fixup! feat(react-icons): decouple griffel styles to follow .styles pattern
