--- packages/react-icons/src/.gitignore
@@ -1,1 +1,0 @@
-fonts/
