--- packages/react-icons/src/utils/fonts/.gitignore
@@ -1,1 +1,0 @@
-FluentSystemIcons*
