+++ packages/react-icons/src/utils/fonts/createFluentFontIcon.shared.ts
@@ -1,0 +1,9 @@
+export const enum FontFile {
+    Filled = 0,
+    Regular = 1,
+    Resizable = 2,
+	Light = 3
+}
+
+
+
